+++ internal/consistency/options_test.go
@@ -3,7 +3,7 @@ package consistency
 import (
 	"testing"
 
-	"github.com/project-kessel/inventory-api/test/helpers"
+	"github.com/project-kessel/inventory-api/internal/helpers"
 	"github.com/spf13/pflag"
 	"github.com/stretchr/testify/assert"
 )
