+++ internal/consumer/options_test.go
@@ -5,7 +5,7 @@ import (
 
 	"github.com/project-kessel/inventory-api/internal/consumer/auth"
 	"github.com/project-kessel/inventory-api/internal/consumer/retry"
-	"github.com/project-kessel/inventory-api/test/helpers"
+	"github.com/project-kessel/inventory-api/internal/helpers"
 	"github.com/spf13/pflag"
 	"github.com/stretchr/testify/assert"
 )
