+++ internal/storage/postgres/options_test.go
@@ -3,7 +3,7 @@ package postgres
 import (
 	"testing"
 
-	"github.com/project-kessel/inventory-api/test/helpers"
+	"github.com/project-kessel/inventory-api/internal/helpers"
 	"github.com/spf13/pflag"
 	"github.com/stretchr/testify/assert"
 )
@@ -31,7 +31,7 @@ func TestOptions_AddFlags(t *testing.T) {
 	prefix := "consumer.postgres"
 	fs := pflag.NewFlagSet("", pflag.ContinueOnError)
 	test.options.AddFlags(fs, prefix)
-	
+
 	helpers.AllOptionsHaveFlags(t, prefix, fs, *test.options, nil)
 }
